requirements.txt
Django==5.1.4
djangorestframework==3.14.0
python-decouple==3.5
cloudinary==1.32.0
cloudinary-storage==1.6.0
mysqlclient==2.1.1
djangorestframework-simplejwt==5.2.0
